split compliance
